restructured peopleList & Profile files in folders
