+++ src/core/rootSaga.js
@@ -1,6 +1,6 @@
 import { watchFetchMovies } from "../features/movies/moviesSaga"
-import { watchFetchPeople } from "../features/people/peopleSaga";
-import { watchFetchProfile } from "../features/people/profileSaga";
+import { watchFetchPeople } from "../features/people/PeopleList/peopleSaga";
+import { watchFetchProfile } from "../features/people/Profile/profileSaga";
 import { all } from "redux-saga/effects";
 
 export default function* rootSaga() {
